Version updates; fixes; linebreak fix

+++ src/install-package.ts
@@ -1,4 +1,4 @@
-import { writeFileSync } from "fs";
+import { writeFileSync, readFileSync, existsSync } from "fs";
 import { resolve } from "path";
 import { mkdir } from "st-mkdir";
 import { executeSync } from "./execute";
@@ -6,6 +6,7 @@ import { getPackageVersionedPath } from "./get-package-versioned-path";
 import { invalidatePackageCachePath } from "./invalidate-package-cache";
 import { osPath } from "./os-path";
 import { safelyResolvePackageCachePath } from "./safely-resolve-package-cache-path";
+import { deletePathOrFile } from "st-rm-rf";
 
 export const latestInstalledVersionCacheFileName = "latest_installed_version";
 
@@ -13,22 +14,35 @@ export const installPackage = async (packageName: string, version: string): Prom
   //invalidatePackageCachePath(packageName);
   const installCachePath = getPackageVersionedPath(packageName, version);
 
+  try {
+    // clean cache
+    deletePathOrFile(installCachePath);
+  } catch (e) { }
+
   mkdir(resolve(installCachePath));
 
-  writeFileSync(
-    resolve(installCachePath, "package.json"),
-    {
-        "name": packageName,
-        "version": version,
-        "description": "auto",
-        "repository": "auto",
-        "license": "ISC"
-    },
-  );
+  const targetPackageJSONPath = resolve(installCachePath, "package.json");
 
   writeFileSync(resolve(installCachePath, "README"), `auto`);
 
   executeSync(["npm", "install", `${packageName}@${version}`, "--prefix", `"${installCachePath}"`]);
+
+  if (!existsSync(targetPackageJSONPath)) {
+
+    writeFileSync(
+      targetPackageJSONPath,
+      JSON.stringify({
+        "name": packageName + '',
+        "version": version,
+        "description": "auto",
+        "repository": "auto",
+        "license": "ISC"
+      }, null, 4),
+      {
+        encoding: 'utf8'
+      }
+    );
+  }
 
   // write version into cache file for offline cache resolve
   writeFileSync(osPath.resolve(safelyResolvePackageCachePath(packageName), latestInstalledVersionCacheFileName), version);
